backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2

# AI services
openai==1.3.7
Pillow==10.1.0
pytesseract==0.3.10

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
respx==0.20.2

# Email validation
email-validator==2.1.0

# Code quality
ruff==0.1.6
black==23.11.0
isort==5.12.0
pre-commit==3.5.0

# Document parsing libraries
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.5.1
matplotlib==3.8.2
numpy==1.25.2
tabulate==0.9.0

# Other utilities
python-dotenv==1.0.0
loguru==0.7.2
